+++ SendBirdSDK.framework/Headers/SBDConstants.h
@@ -14,10 +14,12 @@
 #import "SBDTypes.h"
 
 @class SBDGroupChannel, SBDOpenChannel;
-@class SBDBaseMessage, SBDFileMessage;
+@class SBDBaseMessage, SBDUserMessage, SBDFileMessage;
+@class SBDUser;
 @class SBDError;
 
 typedef void (^SBDOpenChannelHandler)(SBDOpenChannel * _Nullable channel, SBDError * _Nullable error);
+typedef void (^SBDUserMessageHandler)(SBDUserMessage * _Nullable message, SBDError * _Nullable error);
 typedef void (^SBDFileMessageHandler)(SBDFileMessage * _Nullable message, SBDError * _Nullable error);
 typedef void (^SBDGetMessagesHandler)(NSArray<SBDBaseMessage *> * _Nullable messages, SBDError * _Nullable error);
 typedef void (^SBDBinaryProgressHandler)(int64_t bytesSent, int64_t totalBytesSent, int64_t totalBytesExpectedToSend);
@@ -45,5 +47,7 @@ typedef void(^SBDGetPushTokensHandler)(NSArray<NSString *> * _Nonnull pushTokens
 typedef void(^SBDGroupChannelPushTriggerOptionHandler)(SBDGroupChannelPushTriggerOption pushTriggerOption,
                                                        SBDError * _Nullable error);
 typedef void(^SBDErrorHandler)(SBDError * _Nullable error);
+typedef void(^SBDVoidHandler)(void);
+typedef void(^SBDUserHandler)(SBDUser * _Nullable user, SBDError * _Nullable error);
 
 #endif /* SBDConstants_h */