+++ SendBirdSDK.framework/Headers/SBDFileMessage.h
@@ -115,6 +115,16 @@
 @property (strong, nonatomic, readonly, nullable) NSArray<SBDThumbnail *> *thumbnails;
 
 /**
+ *  Represents the dispatch state of the message.
+ *  If message is not dispatched completely to the SendBird server, the value is `SBDMessageRequestStatePending`.
+ *  If failed to send the message, the value is `SBDMessageRequestStateFailed`.
+ *  And if success to send the message, the value is `SBDMessageRequestStateSucceeded`.
+ *
+ *  @since 3.0.141
+ */
+@property (assign, nonatomic, readonly) SBDMessageRequestState requestState;
+
+/**
  *  Builds file message with the information which is releated to file.
  *
  *  @param url        The file URL.