+++ SendBirdSDK.framework/Headers/SBDTypes.h
@@ -113,6 +113,8 @@ typedef NS_ENUM(NSInteger, SBDErrorCode) {
     SBDErrorFileUploadCancelFailed = 800230,
     SBDErrorFileUploadCancelled = 800240,
 	SBDErrorFileUploadTimeout = 800250,
+    SBDErrorTimerWasExpired = 800301,
+    SBDErrorTimerWasAlreadyDone = 800302,
 };
 
 /**
@@ -559,12 +561,28 @@ typedef NS_ENUM(NSUInteger, SBDGroupChannelPushTriggerOption) {
  
  @since 3.0.134
  */
-
 typedef NS_ENUM(NSUInteger, SBDPushTokenType) {
     SBDPushTokenTypeNone = 0,
     SBDPushTokenTypeGCM,
     SBDPushTokenTypeAPNS,
     SBDPushTokenTypeAPNSVoIP,
+};
+
+/**
+ Constants of type to describe message's request state
+ 
+ - SBDMessageRequestStateNone: SHOULD NOT BE. If you got a message instance from SDK, the message can't have this value.
+ - SBDMessageRequestStatePending: Indicates the state of the message returned when trying to send a message. The message with the pending state means that is not dispatched completely to the SendBird server. The pending message should be replaced with a message (failed or succeeded) from the callback.
+ - SBDMessageRequestStateFailed: Indicates the state of the message that failed to send the message.
+ - SBDMessageRequestStateSucceeded: Indicates the state of the message that success to send the message.
+ 
+ @since 3.0.141
+ */
+typedef NS_ENUM(NSUInteger, SBDMessageRequestState) {
+    SBDMessageRequestStateNone = 0,
+    SBDMessageRequestStatePending,
+    SBDMessageRequestStateFailed,
+    SBDMessageRequestStateSucceeded,
 };